Extract DeviceLimiter strategy (#563 Sprint 4, PR 4b)
Complete the RateLimitChecker device split: the two-tier device-fingerprint
matcher, the per-form effective-settings resolver, and the signal-row writer
move to FreeFormCertificate\Security\DeviceLimiter (constructed with the
injected RateLimitSupport). RateLimitChecker::check_device_limit() /
get_device_effective_settings() / record_device_signals() now delegate; the
STRONG_SIGNALS / WEAK_SIGNALS registry stays on RateLimitChecker, and the
RateLimiter facade + all external call sites are unchanged.

DeviceLimiter is keyed by (form_id, signals[]) rather than a string
identifier, so it deliberately does not implement RateLimitStrategy. New
DeviceLimiterTest pins the non-SQL branches + the effective-settings
resolver; the SQL matching path stays covered by RateLimiterTest.
RateLimitChecker drops from ~1119 to ~790 lines across Sprint 4.
