+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format follows [Keep a Changelog] (https://keepachangelog.com/en/1.1.0/).
 - Internal refactor (#563 Sprint 3, PR 3a) — began dismantling the `Core\Utils` god-utility (364 inbound refs) by extracting the filename builders into a focused `FreeFormCertificate\Core\FilenameHelper`: `sanitize_filename()`, `build_pdf_filename()`, and `get_export_filename()`. All call sites repointed (no back-compat wrappers). Pure string helpers, behavior-preserving; the `Core\Utils` request-input accessors follow in PR 3b.
 - Internal refactor (#563 Sprint 3, PR 3b) — extracted the request-input accessors from `Core\Utils` into `FreeFormCertificate\Core\RequestInput`: `get_post_array()`, `get_post_string()`, `get_get_string()`, `get_post_int()`, `get_post_bool()`. All 207 call sites across ~50 files repointed (no back-compat wrappers). Behavior-preserving — typed `$_POST`/`$_GET` readers moved verbatim; nonce verification stays the caller's responsibility. `Core\Utils` drops from ~708 to ~430 lines across phase 1.
 - Internal refactor (#563 Sprint 4, PR 4a) — began splitting the `RateLimitChecker` god-class into per-dimension strategy classes behind a `RateLimitStrategy` interface with an injected `RateLimitSupport` (settings + message formatter): `IpLimiter`, `EmailLimiter`, `CpfLimiter` (all in `FreeFormCertificate\Security\`). `RateLimitChecker::check_{ip,email,cpf}_limit()` now delegate to the strategies; the `RateLimiter` facade and every external call site are unchanged. First sprint to adopt instance-deps + interface (the B2/B3 direction). Behavior-preserving (full suite green); the `DeviceLimiter` follows in PR 4b. Also dropped four now-unused `Core\Utils` imports left in the submission guards after Sprint 3.
+- Internal refactor (#563 Sprint 4, PR 4b) — completes the `RateLimitChecker` device split: the two-tier device-fingerprint matcher, the per-form effective-settings resolver, and the signal-row writer move to `FreeFormCertificate\Security\DeviceLimiter` (constructed with the injected `RateLimitSupport`). `RateLimitChecker::check_device_limit()` / `get_device_effective_settings()` / `record_device_signals()` now delegate; the `STRONG_SIGNALS` / `WEAK_SIGNALS` registry stays on `RateLimitChecker`, and the `RateLimiter` facade + all external call sites are unchanged. `DeviceLimiter` is keyed by `(form_id, signals[])` so it does not implement `RateLimitStrategy`. Behavior-preserving — `RateLimitChecker` drops from ~1119 to ~790 lines across Sprint 4.
 
 ## [6.11.2] (2026-06-21)
